// 0.legacy-frontend/library-app/src/models/MessageModel.ts
class MessageModel {
    title: string;
    question: string;
    id?: number;
    userEmail?: string;
    adminEmail?: string;
    response?: string;
    closed?: boolean;
    _links?: {
        self?: {
            href?: string;
        };
    };

    constructor(title: string, question: string) {
        this.title = title;
        this.question = question;
    }
}
export default MessageModel;